fix(tabs): Accessibility and keyboard interaction fixes
This commit does a few things related to accessibility:
1) Hides md-dummy-tabs completely from screen readers with aria-hidden.
With this change, screen readers do not hear the dummy tabs read out
when using linear navigation. Instead of focusing the dummy tabs, the
visible tabs are focused. There was a comment that mentions focusing the
dummy tabs as a way to avoid animation issues; however, I could not
replicate the mentioned issues.
2) Fixes the ARIA roles so users can hear "tab 1 of 3" instead of just
"tab" by moving role="tablist" to the direct parent element (which
contains the items marked role="tab"). (VoiceOver in particular)
3) Fixes keyboard interaction to match recommended ARIA practices by
focusing on the active tab when using TAB navigation, then resolving the
focus/active tab discrepancy on pressing TAB within the tablist. This
prevents the behavior where tabbing away from the tablist then back
results in the previously focused tab getting focus instead of the
active tab.
4) Note that there is a SCSS change added in to retain current behavior
when the active tab is added to the tab order; however, this should
likely be reconsidered as it actively overrides default focus outline
behavior.
5) Removes unsupported ARIA attribute (aria-activedescendant) from
md-tabs-canvas and removes md-tab-id from dummy tabs to prevent
duplicate setting of aria-controls.
6) Moved id to tab item from dummy tab so the aria-labelledby on the tab
content (role="tabpanel") items is set as per ARIA recommendation.

Tests were updated and fleshed out to check that the active tab is the
only tab allowed in the tab order and that focused/active tabs are
reconciled on tabbing out of the tablist.

Closes #10075.